requirements.txt
python-telegram-bot==20.7
googletrans==4.0.0-rc1
SpeechRecognition
pytesseract
Pillow
ffmpeg-python
